improve nesting objects under arrays
